Extracting models structures

Functions extracting the structure of the models for each package nlme, lme4 and saemix

modelStructnlme(m1, m0)

modelStructlme4(m1, m0, linmodel)

modelStructsaemix(m1, m0)

Arguments

m1
the model under H1

m0
the model under H0

linmodel
(only for modelStructlme4) a boolean to specify whether the model is linear or not

Value

A list with the following components:

detailStruct

a data frame containing 8 variables: name with the name of the model parameters, var1 and var2 with the names of the two variances associated with each covariance parameter, type giving the type of parameter (beta for fixed effects, sd for variances and co for covariances), tested equal to TRUE if the parameter is tested and FALSE otherwise, block giving the block number to which the variance component parameter belongs (equal 0 for fixed effects), covTested indicating whether a covariance is tested without the associated variances being tested, and covInBlock indicating whether a covariance is tested within a block of the complete covariance matrix

dims

a list with the dimensions of the models (nbFE1 and nbFE0 the number of fixed effects in m1 and m0, nbRE1 and nbRE0 the number of random effects in m1 and m0 and dimSigma the number of residual error parameters)

structGamma

the structure of the covariance matrix of the random effects as a list of three logical elements: diag, full and blockDiag, equal to TRUE if the matrix is diagonal, full or block-diagonal respectively.