Monte Carlo approximation of chi-bar-square weights

Approximation of the chi-bar-square weigths via Monte Carlo approximation.

approxWeights(x, df, q)

Arguments

x
a vector of i.i.d. random realizations of the target chi-bar-square distribution

df
a vector containing the degrees of freedom of the chi-squared components

q
the empirical quantile of x used to choose the \(p-2\) values \(c_1, \dots, c_{p-2}\) (see Details)

Value

A vector containing the estimated weights, as well as their covariance matrix.

Details

The chi-bar-square distribution \(\bar{\chi}^2(I,C)\) is a mixture of chi-square distributions. The function provides a method to approximate the weights of the mixture components, when the number of components is known as well as the degrees of freedom of each chi-square distribution in the mixture, and given a vector of simulated values from the target \(\bar{\chi}^2(I,C)\) distribution. Let us assume that there are \(p\) components in the mixture, with degrees of freedom between \(n_1\) and \(n_p\). By definition of a mixture distribution, we have : $$ P(\bar{\chi}^2(I,C) \leq c) = \sum_{i=n_1}^{n_p} w_i P(\chi^2_{i} \leq c)$$ Choosing \(p-2\) values \(c_1, \dots, c_{p-2}\), the function will generate a system of \(p-2\) equations according to the above relationship, and add two additional relationships stating that the sum of all the weights is equal to 1, and that the sum of odd weights and of even weights is equal to 1/2, so that we end up with a system a \(p\) equations with \(p\) variables.

Author

Charlotte Baey <charlotte.baey@univ-lille.fr>