Variance component testing

Perform a likelihood ratio test to test whether a subset of the variances of the random effects are equal to zero. The test is defined by two hypotheses, H0 and H1, and the model under H0 is assumed to be nested within the model under H1.

varTest(
  m1,
  m0,
  control = list(N = 5000, parallel = F, nbcores = 1),
  pval.comp = "bounds",
  fim = "extract"
)

Arguments

m1
a fit of the model under H1, obtained from nlme, lme4 or saemix

m0
a fit of the model under H0, obtained from the same package as m0

control
(optional) a list of control options for the computation of the chi-bar-weights

pval.comp
(optional) the method to be used to compute the p-value, one of: "bounds" (the default), "approx" or "both" (see Details section)

fim
(optional) the method to compute the Fisher Information Matrix. Currently, only fim="extract" is supported.

Value

A list with the following components:

lrt

the likelihood ratio test statistics

ddl

the degrees of freedom of the chi-bar distributions involved in the chi-bar-square distribution

weights

the weights of the limiting chi-bar-square distribution

pval

the p-value of the test

Details

It is possible to tests if any subset of the variances are equal to zero. However, the function does not currently support nested random effects, and assumes that the random effects are Gaussian.

The asymptotic distribution of the likelihood ratio test is a chi-bar-square, with weights that need to be approximated by Monte Carlo methods, apart from some specific cases where they are available explicitly. Therefore, the p-value of the test is not exact but approximated. This computation can be time-consuming, so the default behaviour of the function is to provide bounds on the exact p-value, which can be enough in practice to decide whether to reject or not the null hypothesis. This is triggered by the option pval.comp="bounds". To compute an approximation of the exact p-value, one should use the option pval.comp="approx" or pval.comp="both".

The control argument controls the options for chi-bar-square weights computation. It is a list with the following elements: M the size of the Monte Carlo simulation, parallel a boolean for parallel computing and nbcores the number of cores to be used in case of parallel computing. Default is M=5000, parallel=FALSE and nbcores=1.

References

Baey C, Cournède P-H, Kuhn E, 2019. Asymptotic distribution of likelihood ratio test statistics for variance components in nonlinear mixed effects models. Computational Statistics and Data Analysis 135:107-122.

Silvapulle MJ, Sen PK, 2011. Constrained statistical inference: order, inequality and shape constraints.

Examples

# load nlme package and example dataset
library(nlme)
data(Orthodont)

# fit the two models under H1 and H0
lm1.h1.nlme <- lme(distance ~ 1 + Sex + age + age*Sex, random = ~ 1 + age | Subject, data = Orthodont, method = "ML")
lm1.h0.nlme <- lme(distance ~ 1 + Sex + age + age*Sex, random = ~ 1 | Subject, data = Orthodont, method = "ML")

# compare them (order is important: m1 comes first)
varTest(lm1.h1.nlme,lm1.h0.nlme)

#> Variance components testing in mixed-effects models
#> (models fitted using the nlme package)
#> 
#> Testing that the variance of age is null
#> model under H1: distance ~ 1 + Sex + age + age * Sex (fixed effects) , ~1 + age | Subject (random effects)
#> model under H0: distance ~ 1 + Sex + age + age * Sex (fixed effects) , ~1 | Subject (random effects)
#> 
#> Likelihood ratio test statistics: 
#>  LRT =  0.83311 
#> 
#> Limiting distribution: 
#> mixture of 2 chi-bar-square distributions with degrees of freedom 1, 2 
#> 
#> lower-bound for p-value: 0.51035  upper bound for p-value: 0.51035